Add ueberauth package for oauth sign in
Add ueberauth to the project so we can implement Google sign in. This
will allow the project to use OAuth/social sign in instead of
implementing its own. We can also use this to hook up to Google Fit API.
Google will be the first we implement for that reason. Other 3rd party
that supports OAuth may be implemented at a later time.

Also create an example for dotenv(`.env`) file that will be used for
handling credentials. Currently it is used by the docker test container.